Rename _jpeg module to reflect new generic usage
