+++ image_viewer/image/_read.pyi
